Break out s3 code into other file
